Final Version
Fixed Bugs
Added Quests
Added Instructions
Added End Scene
Added Comments on simple things.
Added rails to the outside.
Added End Game
Gave player visuals for if they passed or failed a quest.
Gave the player more direction.
Changed colors when the player is talking or an NPC is talking to make it easier for the player.